+++ DeathRace/Scene.h
@@ -15,6 +15,6 @@ private:
 	Vector2 sidewalkLeftStartPos, sidewalkLeftEndPos, sidewalkRightStartPos, sidewalkRightEndPos;
 	Player *player1, *player2;
 	KeyboardPlayerInput *player1Input, *player2Input;
-	void DrawDottedLine(Vector2 startPos, Vector2 endPos, int thickness, Color color);
+	void DrawDottedLine(Vector2 startPos, Vector2 endPos, float thickness, Color color);
 };
